src/requirements.txt
python-telegram-bot==12.0.0b1
jsonpickle==1.2